+++ include/aikido/common/detail/algorithm-impl.hpp
@@ -21,7 +21,7 @@ const T& clamp(const T& v, const T& lo, const T& hi, Compare comp)
 template <class T>
 const T& clamp(const T& v, const T& lo, const T& hi)
 {
-  return clamp(v, lo, hi, std::less<T>());
+  return ::aikido::common::clamp(v, lo, hi, std::less<T>());
 }
 
 } // namespace common