+++ src/robot/Robot.cpp
@@ -11,7 +11,6 @@
 #include "aikido/planner/dart/ConfigurationToEndEffectorOffset.hpp"
 #include "aikido/planner/dart/ConfigurationToTSR.hpp"
 #include "aikido/planner/vectorfield/VectorFieldConfigurationToEndEffectorOffsetPlanner.hpp"
-#include "aikido/robot/util.hpp"
 #include "aikido/statespace/GeodesicInterpolator.hpp"
 #include "aikido/statespace/StateSpace.hpp"
 #include "aikido/statespace/dart/MetaSkeletonStateSpace.hpp"
@@ -455,7 +454,7 @@ trajectory::TrajectoryPtr Robot::planToOffset(
         trajPostProcessor) const
 {
   // Get Body Node
-  auto bn = mMetaSkeleton->getBodyNode(bodyNodeName);
+  auto bn = getRootSkeleton()->getBodyNode(bodyNodeName);
   if (!bn)
   {
     dtwarn << "Request body node not present in robot '" << mName << "'"
@@ -483,7 +482,7 @@ trajectory::TrajectoryPtr Robot::planToOffset(
     dartPlanner
         = std::make_shared<VectorFieldConfigurationToEndEffectorOffsetPlanner>(
             mStateSpace,
-            getMetaSkeletonClone(),
+            getMetaSkeleton(),
             vfParams.distanceTolerance,
             vfParams.positionTolerance,
             vfParams.angularTolerance,
@@ -547,7 +546,7 @@ trajectory::TrajectoryPtr Robot::planToTSR(
     const std::string bodyNodeName,
     const constraint::dart::TSRPtr& tsr,
     const constraint::TestablePtr& testableConstraint,
-    std::size_t maxSamples,
+    const util::PlanToTSRParameters& params,
     const std::shared_ptr<planner::ConfigurationToConfigurationPlanner>&
         planner,
     const std::shared_ptr<aikido::planner::TrajectoryPostProcessor>
@@ -555,7 +554,7 @@ trajectory::TrajectoryPtr Robot::planToTSR(
     const distance::ConstConfigurationRankerPtr& ranker) const
 {
   // Get Body Node
-  auto bn = mMetaSkeleton->getBodyNode(bodyNodeName);
+  auto bn = getRootSkeleton()->getBodyNode(bodyNodeName);
   if (!bn)
   {
     dtwarn << "Request body node not present in robot '" << mName << "'"
@@ -568,7 +567,10 @@ trajectory::TrajectoryPtr Robot::planToTSR(
       mStateSpace,
       mStateSpace->getScopedStateFromMetaSkeleton(mMetaSkeleton.get()),
       bn,
-      maxSamples,
+      params.maxSamplingTries,
+      params.batchSize,
+      params.maxBatches,
+      params.numMaxIterations,
       tsr,
       testableConstraint);
 
@@ -623,7 +625,7 @@ trajectory::TrajectoryPtr Robot::planToTSR(
 trajectory::TrajectoryPtr Robot::planToTSR(
     const std::string bodyNodeName,
     const constraint::dart::TSRPtr& tsr,
-    std::size_t maxSamples,
+    const util::PlanToTSRParameters& params,
     const std::shared_ptr<planner::ConfigurationToConfigurationPlanner>&
         planner,
     const std::shared_ptr<aikido::planner::TrajectoryPostProcessor>
@@ -634,7 +636,7 @@ trajectory::TrajectoryPtr Robot::planToTSR(
       bodyNodeName,
       tsr,
       getSelfCollisionConstraint(),
-      maxSamples,
+      params,
       planner,
       trajPostProcessor,
       ranker);